+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-# nf-core/callingcards: Changelog
+# nf-core/callingcards-mammals: Changelog
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.0.0 - [date]
+## v1.0dev - [date]
 
-Initial release of nf-core/callingcards, created with the [nf-core](https://nf-co.re/) template.
+Initial release of nf-core/callingcards-mammals, created with the [nf-core](https://nf-co.re/) template.
 
 ### `Added`